backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-environ>=0.11
drf-spectacular>=0.27
psycopg2-binary>=2.9
gunicorn>=21.2
pytest-django>=4.8
pytest-benchmark>=4.0
reportlab>=4.0
locust>=2.24
